Add process information

+++ src/Model/Table/QueueProcessesTable.php
@@ -71,10 +71,18 @@ class QueueProcessesTable extends Table {
 			'className' => 'Queue.QueuedJobs',
 			'foreignKey' => 'workerkey',
 			'bindingKey' => 'workerkey',
-			'propertyName' => 'active_job',
+			'propertyName' => 'jobs',
 			'conditions' => [
-				'CurrentQueuedJobs.completed IS NULL',
+				'QueuedJobs.completed IS NULL',
 			],
+		]);
+
+
+		$this->hasOne('ActiveQueuedJob', [
+			'className' => 'Queue.QueuedJobs',
+			'foreignKey' => 'id',
+			'bindingKey' => 'active_job_id',
+			'propertyName' => 'active_job'
 		]);
 	}
 
@@ -139,16 +147,30 @@ class QueueProcessesTable extends Table {
 	}
 
 	/**
+	 * Returns the processes for this server in multi-server environment
+	 * otherwise, return all
+	 * @return SelectQuery
+	 */
+	public function findForServer(): SelectQuery {
+		if (Configure::read('Queue.multiserver')) {
+			return $this->find()->where(['server' => $this->buildServerString()]);
+		}
+
+		return $this->find();
+	}
+
+	/**
 	 * @param string $pid
 	 * @param string $key
 	 *
 	 * @return int
 	 */
-	public function add(string $pid, string $key): int {
+	public function add(string $pid, string $key, string $arguments = NULL): int {
 		$data = [
 			'pid' => $pid,
 			'server' => $this->buildServerString(),
 			'workerkey' => $key,
+			'arguments' => $arguments,
 		];
 
 		$queueProcess = $this->newEntity($data);
@@ -164,7 +186,7 @@ class QueueProcessesTable extends Table {
 	 *
 	 * @return void
 	 */
-	public function update(string $pid): void {
+	public function update(string $pid, int $jobId = NULL): void {
 		$conditions = [
 			'pid' => $pid,
 			'server IS' => $this->buildServerString(),
@@ -177,6 +199,8 @@ class QueueProcessesTable extends Table {
 		}
 
 		$queueProcess->modified = new DateTime();
+		$queueProcess->active_job_id = $jobId;
+
 		$this->saveOrFail($queueProcess);
 	}
 
@@ -201,7 +225,31 @@ class QueueProcessesTable extends Table {
 		$timeout = Config::defaultworkertimeout();
 		$thresholdTime = (new DateTime())->subSeconds($timeout);
 
-		return $this->deleteAll(['modified <' => $thresholdTime]);
+		return $this->deleteAll(['modified <' => $thresholdTime]) + $this->cleanNotActiveProcesses();
+	}
+
+	/**
+	 * Remove QueueProcesses which have not an active PID attached
+	 * @return int
+	 */
+	public function cleanNotActiveProcesses(): int {
+		$cleaned_processed = 0;
+		foreach ($this->findForServer() as $process) {
+			if ($this->isProcessRunning($process->pid) === FALSE) {
+				if (!is_null($process->active_job_id)) {
+					$job = $this->CurrentQueuedJobs->findById($process->active_job_id)->first();
+
+					if ($job && is_null($job->completed)) {
+						$this->CurrentQueuedJobs->reset($job->id, TRUE);
+					}
+				}
+
+				$this->delete($process);
+				$cleaned_processed++;
+			}
+		}
+
+		return $cleaned_processed;
 	}
 
 	/**
@@ -302,9 +350,12 @@ class QueueProcessesTable extends Table {
 		if (!$killed) {
 			exec('kill -' . $sig . ' ' . $pid);
 		}
+
 		sleep(1);
 
-		$this->deleteAll(['pid' => $pid]);
+		if ($this->isProcessRunning((int)$pid) !== TRUE) {
+			$this->deleteAll(['pid' => $pid]);
+		};
 	}
 
 	/**
@@ -346,6 +397,25 @@ class QueueProcessesTable extends Table {
 		}
 
 		return $serverName ?: null;
+	}
+
+	/**
+	 * Checks if it is determinable to check if process is running.
+	 * If not, it will return NULL
+	 * @param string $pid
+	 * @return NULL|bool
+	 */
+	public function isProcessRunning (string $pid) : ?bool
+	{
+		if (function_exists('posix_getpgid')) {
+			return posix_getpgid((int)$pid) !== FALSE;
+		}
+
+		if (is_dir('/proc/')) {
+			return file_exists('/proc/' . $pid);
+		}
+
+		return NULL;
 	}
 
 }
